+++ NEWS.md
@@ -2,8 +2,9 @@ Version 0.0.1.999
 =================
 
 * [#3](https://github.com/vspinu/mlvocab/issues/3) Add support for character vector corpus.
-* Add support for `data.frame` corpus 
-* New index generator function `tidf` to produce flat indices as `data.frames`.
+* Add support for `data.frame` corpus.
+* New index generator function `tix_df` to produce flat indices as `data.frames`.
+* Rename `tixyz` function into `tix_xyz`.
 
 
 Version 0.0.1